+++ CHANGELOG.md
@@ -4,3 +4,5 @@
 - PinField, source: Updated keyboard tests
 - Flyout: Correctly calculate position for scaled down triggers
 - Flyout: Using vanilla event listeners on trigger, you can pass additional custom handlers yourself
+- View: Responsive textAlign support
+- Toast: Warning color support